# Use selective checkpointer for conversations