@kdid 2025-03-13
@mbklein - I'm encountering this error (not 100% of the time but it's happened more than once) when I ask follow up questions with the same ref. Not sure if this is related to my recent work or the interaction of it with this PR.....

```
[ERROR] EventStreamError: An error occurred (validationException) when calling the InvokeModelWithResponseStream operation: messages.4: Did not find 1 `tool_result` block(s) at the beginning of this message. Messages following `tool_use` blocks must begin with a matching number of `tool_result` blocks.
Traceback (most recent call last):
  File "/opt/python/honeybadger/contrib/aws_lambda.py", line 91, in wrapped_handler
    reraise(*exc_info)
  File "/opt/python/honeybadger/contrib/aws_lambda.py", line 49, in reraise
    raise value
  File "/opt/python/honeybadger/contrib/aws_lambda.py", line 83, in wrapped_handler
    return handler(aws_event, aws_context, *args, **kwargs)
  File "/var/task/handlers.py", line 78, in chat
    raise e
  File "/var/task/handlers.py", line 72, in chat
    search_agent.invoke(config.question, config.ref, forget=config.forget, docs=config.docs, callbacks=callbacks)
  File "/var/task/agent/search_agent.py", line 101, in invoke
    return self.search_agent.invoke(
  File "/opt/python/langgraph/pregel/__init__.py", line 2375, in invoke
    for chunk in self.stream(
  File "/opt/python/langgraph/pregel/__init__.py", line 2031, in stream
    for _ in runner.tick(
  File "/opt/python/langgraph/pregel/runner.py", line 230, in tick
    run_with_retry(
  File "/opt/python/langgraph/pregel/retry.py", line 40, in run_with_retry
    return task.proc.invoke(task.input, config)
  File "/opt/python/langgraph/utils/runnable.py", line 546, in invoke
    input = step.invoke(input, config, **kwargs)
  File "/opt/python/langgraph/utils/runnable.py", line 310, in invoke
    ret = context.run(self.func, *args, **kwargs)
  File "/var/task/agent/search_agent.py", line 40, in call_model
    response: BaseMessage = self.model.invoke(messages)
  File "/opt/python/langchain_core/runnables/base.py", line 5365, in invoke
    return self.bound.invoke(
  File "/opt/python/langchain_core/language_models/chat_models.py", line 307, in invoke
    self.generate_prompt(
  File "/opt/python/langchain_core/language_models/chat_models.py", line 843, in generate_prompt
    return self.generate(prompt_messages, stop=stop, callbacks=callbacks, **kwargs)
  File "/opt/python/langchain_core/language_models/chat_models.py", line 683, in generate
    self._generate_with_cache(
  File "/opt/python/langchain_core/language_models/chat_models.py", line 908, in _generate_with_cache
    result = self._generate(
  File "/opt/python/langchain_aws/chat_models/bedrock.py", line 652, in _generate
    return generate_from_stream(stream_iter)
  File "/opt/python/langchain_core/language_models/chat_models.py", line 91, in generate_from_stream
    generation = next(stream, None)
  File "/opt/python/langchain_aws/chat_models/bedrock.py", line 589, in _stream
    for chunk in self._prepare_input_and_invoke_stream(
  File "/opt/python/langchain_aws/llms/bedrock.py", line 1099, in _prepare_input_and_invoke_stream
    for chunk in LLMInputOutputAdapter.prepare_output_stream(
  File "/opt/python/langchain_aws/llms/bedrock.py", line 478, in prepare_output_stream
    for event in stream:
  File "/opt/python/botocore/eventstream.py", line 592, in __iter__
    parsed_event = self._parse_event(event)
  File "/opt/python/botocore/eventstream.py", line 608, in _parse_event
    raise EventStreamError(parsed_response, self._operation_name)
 ```